VideoCommon: Move imgui-related functions/data to ImGuiManager class
